Move CMPTradeList into seperate file
